# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: SorobanPay Backend API
  description: |
    REST API for the SorobanPay recurring-payment protocol.

    Merchants authenticate via a SEP-10 style challenge-response flow:
    1. `GET /api/v1/auth/challenge?account=G…` — obtain unsigned challenge XDR
    2. Sign the returned transaction with Freighter (or your private key)
    3. `POST /api/v1/auth/token { transaction: "<signed-XDR>" }` — exchange for JWT
    4. Include `Authorization: Bearer <token>` on all protected endpoints

    Rate limit: **100 requests / minute per IP** (HTTP 429 on breach).

    ## Structured error codes (#1074)

    All 4xx/5xx responses return an `ApiError` envelope:
    ```json
    {
      "error": "Human-readable message safe for clients",
      "code": "VALIDATION_ERROR",
      "correlationId": "req_01HXK8MZAB",
      "details": null
    }
    ```
    Stable machine-readable `code` values: `VALIDATION_ERROR`, `UNAUTHORIZED`,
    `FORBIDDEN`, `NOT_FOUND`, `RATE_LIMITED`, `INTERNAL_ERROR`,
    `DEPENDENCY_UNAVAILABLE`, `IDEMPOTENCY_CONFLICT`.
    Use `correlationId` when filing support requests or tracing logs.
  version: 1.0.0
  contact:
    name: SorobanPay
    url: https://github.com/Chrisland58/SorobanPay
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.sorobanpay.example.com
    description: Production (placeholder)

# ─── Security ──────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by `POST /api/v1/auth/token`. Include as `Authorization: Bearer <token>`.
        Tokens expire after 24 hours.
  schemas:
...
    Address:
      type: string
      pattern: "^G[A-Z2-7]{55}$"
      example: "GABC1234MERCHANT5678EXAMPLE000000000000000000000000000XYZ"
    IsoDate:
      type: string
      format: date
      example: "2024-01-15"

    # ── Subscription ──────────────────────────────────────────────────────────
    Subscription:
      type: object
      required: [subscriber, merchant, token, amount, status]
      properties:
        subscriber:
          $ref: "#/components/schemas/Address"
        merchant:
          $ref: "#/components/schemas/Address"
        token:
          $ref: "#/components/schemas/Address"
        amount:
          type: string
          description: Payment amount in stroops (as string to handle big integers)
          example: "10000000"
        status:
          type: string
          enum: [ACTIVE, PAUSED, OVERDUE, CANCELLED, EXPIRED]
        interval:
          type: integer
          nullable: true
          description: Payment interval in seconds (null if not stored in event index)
        nextPaymentDue:
          type: integer
          nullable: true
          description: Unix timestamp of next payment due (null if not computable from index)
        lastPaymentAt:
          type: string
          nullable: true
          description: Ledger timestamp of last executed payment (as string)
        createdAt:
          type: string
          nullable: true
          description: Ledger timestamp when subscription was created

    # ── Payment event ─────────────────────────────────────────────────────────
    Payment:
      type: object
      required: [id, subscriber, merchant, token, amount, ledgerTimestamp, type]
      properties:
        id:
          type: integer
        subscriber:
          $ref: "#/components/schemas/Address"
        merchant:
          $ref: "#/components/schemas/Address"
        token:
          $ref: "#/components/schemas/Address"
        amount:
          type: string
          example: "10000000"
        type:
          type: string
          enum: [executed, payment_transfer_failure]
        ledgerTimestamp:
          type: string
          description: Unix timestamp in seconds (as string)
        createdAt:
          type: string
          format: date-time

    # ── Analytics ─────────────────────────────────────────────────────────────
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Structured API error envelope (#1074) ─────────────────────────────────
    ApiError:
      type: object
      required: [error, code, correlationId]
      properties:
        error:
          type: string
          description: Human-readable error message safe for clients
          example: "Merchant address is missing"
        code:
          type: string
          description: Stable machine-readable error code
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - RATE_LIMITED
            - INTERNAL_ERROR
            - DEPENDENCY_UNAVAILABLE
            - IDEMPOTENCY_CONFLICT
          example: "VALIDATION_ERROR"
        correlationId:
          type: string
          description: Request correlation ID for log tracing — include when filing support requests
          example: "req_01HXK8MZAB"
        details:
          type: object
          nullable: true
          description: Optional structured details (field errors, retry hints, existing request ID)
          additionalProperties: true
    PaginatedMeta:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /api/v1/subscriptions:
    get:
      operationId: listSubscriptions
      summary: List active subscriptions for a merchant
      tags: [Subscriptions]
      security:
        - bearerAuth: []
      parameters:
        - name: merchant
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Address"
          description: Merchant Stellar address
      responses:
        "200":
          description: Array of subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
        "400":
          description: Missing merchant parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/v1/subscriptions/{subscriber}/{merchant}:
    get:
      operationId: getSubscription
      summary: Get a single subscription by subscriber and merchant
      tags: [Subscriptions]
      security:
        - bearerAuth: []
      parameters:
        - name: subscriber
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Address"
        - name: merchant
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Address"
      responses:
        "200":
          description: Subscription detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "404":
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Payments ────────────────────────────────────────────────────────────────
  /api/v1/payments:
    get:
      operationId: listPayments
      summary: List payment history for a merchant
      tags: [Payments]
      security:
        - bearerAuth: []
      parameters:
        - name: merchant
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Address"
        - name: from
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/IsoDate"
          description: Start date (inclusive), ISO 8601
        - name: to
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/IsoDate"
          description: End date (inclusive), ISO 8601
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated payment history
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Payment"
                  meta:
                    $ref: "#/components/schemas/PaginatedMeta"
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized

  # ── Analytics ────────────────────────────────────────────────────────────────
...